requirement.txt
Django==3.2.10
mysqlclient==2.1.0
drf-yasg==1.20.0
django-cors-headers==3.11.0
djangorestframework-simplejwt==5.1.0
django-filter==21.1
user-agents==2.2.0
pypinyin==0.46.0
six==1.16.0
django-simple-captcha==0.5.17